\name{month.name.jp}
\alias{month.name.jp}
\docType{data}
\title{The Japanese name of months}
\description{
	The traditional names of months in Japanese.
}
\usage{data(month.name.jp)}
\format{character vector}
\details{
\code{month.name.jp} is a constant including the Japanese name of months
}
% \source{
% %%  ~~ reference to a publication or URL from which the data were obtained ~~
% }
% \references{
% %%  ~~ possibly secondary sources and usages ~~
% }
\examples{
data(month.name.jp)
month.name.jp[which(month.name=="April")]
}
\keyword{datasets}
