\name{Nippon-deprecated}
\alias{Nippon-deprecated}
\alias{JapanPrefecturesMap}
\title{Deprecated function(s) in the Nippon package}

\description{
Some functions are defunct in the Nippon package. See Details.
}
\section{Details}{ \code{JapanPrefecturesMap} is deprecated. Please use
	\code{JapanPrefMap} in \pkg{NipponMap} package, which is rewrite with
	more modern \pkg{sf} package. \code{JapanPrefecturesMap} in this
	package is defunct.  }
