\name{Nippon-internal}
\alias{Nippon-internal}
\alias{zenkaku}
%\docType{data}
\title{an internal object}
\description{
an internal object in Nippon package
}
\usage{zenkaku}
\format{character vector}
\details{
An internal object in Nippon package: \code{zenkaku}
}
% \source{
% %%  ~~ reference to a publication or URL from which the data were obtained ~~
% }
% \references{
% %%  ~~ possibly secondary sources and usages ~~
% }
% \examples{
% data(zenkaku)
% zenkaku
% }
% \keyword{datasets}
