\name{Nippon-package}
\alias{Nippon-package}
\alias{Nippon}
\docType{package}
\title{Japanese Utility Functions and Data}
\description{
Japan-specific data is sometimes too unhandy for R users to manage. The
utility functions and data in this package disencumber us from such an
unnecessary burden.
}
\details{
\tabular{ll}{
Package: \tab Nippon\cr
Type: \tab Package\cr
Version: \tab 0.6.4\cr
Date: \tab 2017-11-22\cr
License: \tab GPL (>=2)\cr
}
To avoid troubles with Japanese strings, non-ASCII Japanese characters
in R object is sanitized or converted into ASCII characters by utility
functions in this package. Some common data for Japanese are planned to
preset, for preventing from burdensome input. (Those will be reinforced
in further version).
}

\author{Susumu Tanimura \email{aruminat@gmail.com}}

% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
%\keyword{packagelibrary}
%\keyword{character}
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
