\name{Nippon-package}
\alias{Nippon-package}
\alias{Nippon}
\docType{package}
\title{Japanese utility functions and data}
\description{
This package provides some Japanese utility functions and data, which
may be useful for Japanese R users.
}
\details{
\tabular{ll}{
Package: \tab Nippon\cr
Type: \tab Package\cr
Version: \tab 0.1.3\cr
Date: \tab 2011-02-18\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
To avoid troubles with Japanese strings, non-ASCII Japanese characters
in R object is sanitized or converted into ASCII characters by utility
functions in this package. Some common data for Japanese are planned to
preset, for preventing from burdensome input. (Those will be reinforced
in further version).
}

\author{Susumu Tanimura \email{aruminat@gmail.com}}

% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
%\keyword{packagelibrary}
%\keyword{character}
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
